Chore: Try to get walk working
