+++ packages/parser-javascript/tests/fixtures/example.jsx
@@ -1,0 +1,3 @@
+let circle = document.createElement('circle');
+
+let jsx = <button>Test</button>;
